+++ CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+- New command to validate a published app
+
 ## [2.78.6] - 2019-12-04
 ### Refactor
 - Make `vtex init` never change the cloned app's manifest.